# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SpinCash API",
        "description": "SpinCash API Demo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "login",
                "requestBody": {
                    "description": "Data for login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthSuccess"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Register",
                "operationId": "register",
                "requestBody": {
                    "description": "Data for register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthSuccess"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user information",
                "description": "Get user information",
                "operationId": "getUserInfo",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "Login Request",
                "properties": {
                    "email": {
                        "title": "Email",
                        "type": "string",
                        "example": "admin@admin.com"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string",
                        "example": "admin123"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Register Request",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string",
                        "example": "Admin"
                    },
                    "email": {
                        "title": "Email",
                        "type": "string",
                        "example": "admin@admin.com"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string",
                        "example": "admin123"
                    }
                },
                "type": "object"
            },
            "Unauthorized": {
                "title": "Unauthorized",
                "properties": {
                    "message": {
                        "title": "Message",
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
            "AuthSuccess": {
                "title": "Auth Success",
                "description": "Auth success",
                "properties": {
                    "access_token": {
                        "title": "access_token",
                        "type": "string",
                        "example": "1|hAYuajUAUhskjNAYUiuAN"
                    },
                    "token_type": {
                        "title": "token_type",
                        "type": "string",
                        "example": "Bearer"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AuthSuccess"
                }
            },
            "AuthError": {
                "title": "Auth Error",
                "description": "Auth error",
                "properties": {
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
            "UserInfo": {
                "title": "User Information",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string",
                        "example": "Admin"
                    },
                    "email": {
                        "title": "Email",
                        "type": "string",
                        "example": "admin@admin.com"
                    },
                    "tickets": {
                        "title": "Tickets",
                        "type": "integer",
                        "example": "2000"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}